common makeStraightEdges: remove unnecessary cast
This should have been removed in a02eaec8af5a200bd04ee15c07f53108aa102438.

+++ lib/common/routespl.c
@@ -1030,7 +1030,7 @@ void makeStraightEdges(graph_t *g, edge_t **edge_list, size_t e_cnt, int et,
 	    Ppoint_t pts[] = {dumber[0], dumber[1], dumber[2], dumber[3]};
 	    Ppolyline_t spl, line = {.pn = 4, .ps = pts};
 	    make_polyline (line, &spl);
-	    clip_and_install(e0, aghead(e0), spl.ps, (size_t)spl.pn, sinfo);
+	    clip_and_install(e0, aghead(e0), spl.ps, spl.pn, sinfo);
 	}
 	else
 	    clip_and_install(e0, aghead(e0), dumber, 4, sinfo);
